refactor: move delta to workflow_ executor & state
