+++ src/autora/workflow/state/__init__.py
@@ -1,4 +1,3 @@
-from .delta import BaseDelta, BaseState
 from .history import History
 from .param import resolve_state_params
 from .snapshot import Snapshot
